Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [345ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [1.2s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1535  0.8217 -2.2862 -1.8820  0.4123  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.1901 0.1411 0.1661 0.2803 0.1468 0.1155 0.1044 0.0922 0.2275 0.2590
#> [11] 0.1960 0.2666 0.1579 0.2137 0.2102 0.1967 0.2825 0.1776 0.1515 0.1024
#> [21] 0.3174 0.2436 0.2432 0.1987 0.3920 0.1464 0.1562 0.4005 0.2106 0.2024
#> [31] 0.2185 0.2012 0.1725 0.2698 0.2297 0.1525 0.1915 0.3685 0.2288 0.2667
#> [41] 0.1854 0.2051 0.2172 0.1117 0.1099 0.1238 0.2229 0.2237 0.1350 0.1469
#> [51] 0.2519 0.1733 0.2257 0.2180 0.2458 0.2018 0.2458 0.2497 0.2820 0.2146
#> [61] 0.2134 0.1486 0.2360 0.1703 0.1209 0.1698 0.1522 0.3616 0.2245 0.0333
#> [71] 0.2738 0.1573 0.2033 0.1433 0.3145
#> 
#> $betaw
#>  [1] 0.6915 0.6586 0.6272 0.8864 0.5539 0.3803 0.5216 0.5705 0.7835 0.8988
#> [11] 0.5832 0.9249 0.6003 0.7311 0.7636 0.7106 0.7621 0.6822 0.6813 0.6755
#> [21] 0.8822 0.8739 0.7912 0.7177 0.8705 0.6405 0.5694 0.9534 0.7473 0.7643
#> [31] 0.7703 0.6788 0.7228 0.9280 0.7677 0.5964 0.6778 0.8469 0.7209 0.8173
#> [41] 0.6522 0.6982 0.7300 0.6359 0.5525 0.4960 0.7234 0.7616 0.6584 0.6213
#> [51] 0.8763 0.6648 0.7496 0.7415 0.8834 0.6646 0.8529 0.7700 0.8375 0.7107
#> [61] 0.7186 0.6516 0.7338 0.5916 0.6145 0.6614 0.5453 0.8970 0.7330 0.5154
#> [71] 0.7686 0.6986 0.6797 0.4699 0.9723
#>