Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 3.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.1.
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [2ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [2.4s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1606  0.8264 -2.1411 -1.8074  0.0316  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.2218 0.1388 0.1952 0.2820 0.1480 0.1925 0.1057 0.1034 0.2116 0.2260
#> [11] 0.1907 0.2102 0.1969 0.1927 0.2193 0.1856 0.2866 0.1764 0.1462 0.1002
#> [21] 0.3189 0.2078 0.2179 0.2184 0.4009 0.1681 0.1966 0.4154 0.2282 0.2043
#> [31] 0.2170 0.1796 0.1798 0.2238 0.2186 0.1673 0.1855 0.3801 0.2267 0.2774
#> [41] 0.2060 0.2032 0.2169 0.1124 0.1396 0.1360 0.2173 0.2018 0.1338 0.1596
#> [51] 0.2391 0.1707 0.2274 0.2224 0.2254 0.1909 0.2125 0.2532 0.2902 0.2123
#> [61] 0.2110 0.1461 0.2372 0.1911 0.1106 0.1694 0.1754 0.3737 0.2255 0.0366
#> [71] 0.2783 0.1578 0.2063 0.1851 0.3194
#> 
#> $betaw
#>  [1] 0.6900 0.6642 0.6228 0.8849 0.5530 0.3694 0.5199 0.5543 0.7865 0.9063
#> [11] 0.5832 0.9292 0.5902 0.7423 0.7636 0.7153 0.7220 0.6851 0.7171 0.7027
#> [21] 0.8805 0.8747 0.8117 0.7176 0.8354 0.6269 0.5612 0.9423 0.7082 0.7642
#> [31] 0.7710 0.7068 0.6927 0.9281 0.7747 0.5864 0.6780 0.8001 0.7421 0.8049
#> [41] 0.6453 0.6993 0.7301 0.6335 0.5196 0.4877 0.7369 0.7704 0.6615 0.6092
#> [51] 0.8808 0.6678 0.7468 0.7388 0.8859 0.6705 0.8575 0.7602 0.8271 0.7108
#> [61] 0.7379 0.6573 0.7270 0.5880 0.6347 0.6619 0.5371 0.8798 0.7100 0.5042
#> [71] 0.7400 0.6812 0.6795 0.4650 0.9711
#>