Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [266ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [1.2s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1535  0.8217 -2.2862 -1.8820  0.4123  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.2057 0.1447 0.1814 0.2849 0.1325 0.1178 0.1020 0.0912 0.2364 0.2700
#> [11] 0.1737 0.2653 0.1532 0.2208 0.2135 0.1910 0.2829 0.1734 0.1519 0.1040
#> [21] 0.3066 0.2337 0.2417 0.2044 0.3956 0.1498 0.1610 0.3945 0.2085 0.2316
#> [31] 0.2261 0.1854 0.1789 0.2356 0.2217 0.1457 0.2162 0.3687 0.2288 0.2663
#> [41] 0.1826 0.1963 0.2179 0.1123 0.1126 0.1183 0.2138 0.2336 0.1505 0.1472
#> [51] 0.2812 0.1635 0.2286 0.2130 0.2601 0.1843 0.2434 0.2521 0.2779 0.2188
#> [61] 0.2133 0.1464 0.2313 0.1582 0.1258 0.1693 0.1561 0.3591 0.2233 0.0390
#> [71] 0.2731 0.1578 0.1851 0.1411 0.3148
#> 
#> $betaw
#>  [1] 0.6907 0.6495 0.6249 0.8824 0.5640 0.3800 0.5250 0.5719 0.7818 0.8963
#> [11] 0.5834 0.9250 0.6015 0.7273 0.7636 0.7130 0.7581 0.6920 0.6789 0.6551
#> [21] 0.8942 0.8741 0.7924 0.7177 0.8565 0.6384 0.5684 0.9579 0.7519 0.7619
#> [31] 0.7669 0.6993 0.6965 0.9281 0.7727 0.6010 0.6774 0.8463 0.7204 0.8177
#> [41] 0.6531 0.7030 0.7299 0.6336 0.5496 0.4997 0.7454 0.7576 0.6171 0.6210
#> [51] 0.8658 0.6759 0.7449 0.7445 0.8816 0.6741 0.8532 0.7632 0.8428 0.7106
#> [61] 0.7191 0.6567 0.7595 0.5937 0.6050 0.6620 0.5440 0.9006 0.7569 0.4964
#> [71] 0.7731 0.6818 0.6807 0.4701 0.9722
#>