Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [279ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [2.2s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1535  0.8217 -2.2862 -1.8820  0.4123  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.2020 0.1525 0.1847 0.3015 0.1451 0.1305 0.1032 0.1049 0.2257 0.2573
#> [11] 0.1739 0.2666 0.1807 0.2188 0.2216 0.2027 0.2823 0.1708 0.1516 0.1034
#> [21] 0.3100 0.2519 0.2244 0.1902 0.3967 0.1640 0.1611 0.4006 0.2095 0.2250
#> [31] 0.2299 0.1868 0.1787 0.2533 0.2187 0.1319 0.2089 0.3702 0.2289 0.2615
#> [41] 0.1742 0.2060 0.2139 0.1073 0.1107 0.1138 0.2183 0.2106 0.1368 0.1603
#> [51] 0.2802 0.1698 0.2216 0.2154 0.2652 0.1792 0.2494 0.2504 0.2840 0.2020
#> [61] 0.2131 0.1430 0.2351 0.1849 0.1214 0.1662 0.1484 0.3582 0.2250 0.0398
#> [71] 0.2755 0.1577 0.1990 0.1543 0.3136
#> 
#> $betaw
#>  [1] 0.6909 0.6300 0.6244 0.8680 0.5551 0.3782 0.5233 0.5520 0.7838 0.8992
#> [11] 0.5834 0.9249 0.5944 0.7283 0.7636 0.7081 0.7644 0.6981 0.6805 0.6627
#> [21] 0.8905 0.8737 0.8064 0.7178 0.8522 0.6294 0.5684 0.9533 0.7498 0.7624
#> [31] 0.7651 0.6975 0.6972 0.9280 0.7746 0.6103 0.6775 0.8399 0.7195 0.8233
#> [41] 0.6559 0.6977 0.7308 0.6508 0.5517 0.5027 0.7344 0.7668 0.6537 0.6086
#> [51] 0.8662 0.6688 0.7563 0.7431 0.8810 0.6769 0.8524 0.7680 0.8349 0.7111
#> [61] 0.7207 0.6643 0.7384 0.5891 0.6135 0.6657 0.5467 0.9018 0.7209 0.4936
#> [71] 0.7578 0.6829 0.6799 0.4686 0.9725
#>