Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [346ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [3s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1535  0.8217 -2.2862 -1.8820  0.4123  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.1977 0.1425 0.1715 0.2963 0.1453 0.1418 0.0995 0.1052 0.2350 0.2571
#> [11] 0.1734 0.2607 0.1796 0.2070 0.2005 0.1979 0.2821 0.1781 0.1501 0.1025
#> [21] 0.3198 0.2221 0.2415 0.2037 0.3991 0.1640 0.1691 0.3967 0.2188 0.2166
#> [31] 0.2273 0.1837 0.1767 0.2392 0.2240 0.1627 0.1867 0.3700 0.2277 0.2685
#> [41] 0.1909 0.1921 0.2126 0.1101 0.1173 0.1091 0.2123 0.2236 0.1388 0.1552
#> [51] 0.2531 0.1735 0.2212 0.2139 0.2459 0.1758 0.2402 0.2487 0.2879 0.2024
#> [61] 0.2132 0.1406 0.2376 0.1628 0.1182 0.1635 0.1524 0.3678 0.2239 0.0411
#> [71] 0.2750 0.1576 0.1855 0.1587 0.3036
#> 
#> $betaw
#>  [1] 0.6911 0.6550 0.6264 0.8725 0.5549 0.3766 0.5283 0.5517 0.7820 0.8992
#> [11] 0.5834 0.9253 0.5947 0.7346 0.7636 0.7101 0.7662 0.6810 0.6906 0.6734
#> [21] 0.8795 0.8744 0.7926 0.7177 0.8427 0.6295 0.5668 0.9562 0.7290 0.7631
#> [31] 0.7663 0.7015 0.7053 0.9281 0.7713 0.5895 0.6779 0.8410 0.7318 0.8152
#> [41] 0.6503 0.7053 0.7311 0.6412 0.5444 0.5059 0.7491 0.7616 0.6482 0.6134
#> [51] 0.8758 0.6646 0.7569 0.7440 0.8834 0.6788 0.8536 0.7727 0.8300 0.7110
#> [61] 0.7203 0.6697 0.7250 0.5929 0.6198 0.6688 0.5453 0.8881 0.7454 0.4893
#> [71] 0.7608 0.6890 0.6806 0.4681 0.9749
#>