Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 3.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.1.
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [1ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [2.9s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1606  0.8264 -2.1411 -1.8074  0.0316  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.1955 0.1410 0.1909 0.2873 0.1470 0.1756 0.1045 0.1048 0.2223 0.2251
#> [11] 0.2004 0.2241 0.1846 0.2093 0.2090 0.1863 0.2865 0.1710 0.1491 0.1004
#> [21] 0.3243 0.2142 0.2287 0.1998 0.4056 0.1842 0.1962 0.4281 0.2229 0.2145
#> [31] 0.2218 0.1949 0.1755 0.2049 0.2236 0.1611 0.1809 0.3801 0.2328 0.2676
#> [41] 0.1987 0.2176 0.2186 0.1112 0.1306 0.1331 0.2196 0.1961 0.1308 0.1712
#> [51] 0.2330 0.1673 0.2364 0.2200 0.2134 0.1881 0.2148 0.2626 0.2932 0.2055
#> [61] 0.2144 0.1446 0.2394 0.1971 0.1271 0.1780 0.1626 0.3858 0.2251 0.0408
#> [71] 0.2846 0.1572 0.2055 0.1855 0.3001
#> 
#> $betaw
#>  [1] 0.6912 0.6587 0.6234 0.8803 0.5538 0.3718 0.5215 0.5522 0.7845 0.9065
#> [11] 0.5831 0.9282 0.5934 0.7334 0.7636 0.7150 0.7225 0.6978 0.6978 0.7004
#> [21] 0.8745 0.8746 0.8030 0.7177 0.8173 0.6168 0.5613 0.9328 0.7201 0.7633
#> [31] 0.7688 0.6870 0.7103 0.9281 0.7716 0.5906 0.6780 0.7999 0.6798 0.8162
#> [41] 0.6478 0.6914 0.7297 0.6374 0.5296 0.4897 0.7313 0.7727 0.6696 0.5982
#> [51] 0.8830 0.6716 0.7322 0.7402 0.8874 0.6720 0.8571 0.7341 0.8233 0.7110
#> [61] 0.7107 0.6606 0.7153 0.5869 0.6025 0.6519 0.5417 0.8626 0.7197 0.4902
#> [71] 0.6992 0.7017 0.6796 0.4650 0.9758
#>