Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

eiread(ei.object, ...)

Arguments

ei.object
An ei object from the function ei.

...
A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO
References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> i Running 2x2 ei
#>                  
#> i Maximizing likelihood
#> i Running 2x2 ei
#> v Running 2x2 ei [247ms]
#> 
#> \ Beginning importance sampling.
#> v Beginning importance sampling. [1.7s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1535  0.8217 -2.2862 -1.8820  0.4123  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.2032 0.1431 0.1891 0.2884 0.1355 0.1352 0.1040 0.1065 0.2259 0.2720
#> [11] 0.1901 0.2552 0.1596 0.2162 0.2230 0.1981 0.2808 0.1738 0.1520 0.1046
#> [21] 0.3181 0.2472 0.2329 0.1937 0.3977 0.1520 0.1630 0.3994 0.2003 0.2113
#> [31] 0.2251 0.1834 0.1768 0.2726 0.2186 0.1567 0.2048 0.3741 0.2268 0.2622
#> [41] 0.1881 0.2003 0.2126 0.1126 0.1122 0.1139 0.2128 0.2350 0.1356 0.1485
#> [51] 0.2569 0.1741 0.2265 0.2141 0.2635 0.1745 0.2468 0.2425 0.2827 0.2070
#> [61] 0.2130 0.1457 0.2385 0.1829 0.1321 0.1774 0.1474 0.3623 0.2250 0.0463
#> [71] 0.2695 0.1572 0.1812 0.1450 0.3225
#> 
#> $betaw
#>  [1] 0.6908 0.6534 0.6237 0.8794 0.5619 0.3775 0.5223 0.5498 0.7838 0.8958
#> [11] 0.5832 0.9258 0.5999 0.7297 0.7636 0.7100 0.7784 0.6912 0.6783 0.6471
#> [21] 0.8814 0.8738 0.7996 0.7178 0.8482 0.6370 0.5680 0.9542 0.7701 0.7636
#> [31] 0.7673 0.7018 0.7052 0.9280 0.7747 0.5935 0.6776 0.8242 0.7412 0.8225
#> [41] 0.6513 0.7008 0.7311 0.6327 0.5501 0.5027 0.7477 0.7570 0.6569 0.6197
#> [51] 0.8745 0.6639 0.7483 0.7438 0.8812 0.6795 0.8527 0.7900 0.8366 0.7109
#> [61] 0.7218 0.6583 0.7203 0.5894 0.5927 0.6525 0.5470 0.8960 0.7207 0.4720
#> [71] 0.7960 0.7027 0.6809 0.4697 0.9703
#>