Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

Usage

eiread(ei.object, ...)

Arguments

ei.object

An ei object from the function ei.

...

A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO

References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample_ei)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample_ei)
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 3.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 5.
#> ℹ Running 2x2 ei
#> ℹ Maximizing likelihood for `erho` = 0.1.
#> ℹ Running 2x2 ei
#> ✔ Running 2x2 ei [2ms]
#> 
#> ⠙ Beginning importance sampling.
#> ✔ Beginning importance sampling. [1.6s]
#> 
eiread(dbuf, "phi")
#> [1] -1.1606  0.8264 -2.1411 -1.8074  0.0316  0.0000  0.0000
eiread(dbuf, "betab", "betaw")
#> $betab
#>  [1] 0.1951 0.1323 0.1967 0.2921 0.1646 0.1978 0.1141 0.1028 0.2278 0.2333
#> [11] 0.1955 0.2094 0.1956 0.2111 0.1959 0.1837 0.2840 0.1628 0.1523 0.1028
#> [21] 0.3366 0.2120 0.2417 0.2081 0.4138 0.1783 0.1957 0.4224 0.2156 0.2306
#> [31] 0.2020 0.1801 0.1819 0.2440 0.2261 0.1631 0.2152 0.3872 0.2302 0.2736
#> [41] 0.2141 0.2094 0.2017 0.1038 0.1147 0.1543 0.2258 0.2156 0.1302 0.1636
#> [51] 0.2456 0.1808 0.2325 0.2157 0.2201 0.2054 0.2049 0.2606 0.2971 0.1901
#> [61] 0.2127 0.1343 0.2352 0.1926 0.1234 0.1670 0.1771 0.3908 0.2257 0.0421
#> [71] 0.2774 0.1568 0.2092 0.1872 0.3086
#> 
#> $betaw
#>  [1] 0.6912 0.6805 0.6225 0.8762 0.5413 0.3687 0.5086 0.5551 0.7834 0.9046
#> [11] 0.5832 0.9293 0.5906 0.7324 0.7636 0.7161 0.7474 0.7168 0.6758 0.6700
#> [21] 0.8607 0.8746 0.7924 0.7177 0.7847 0.6205 0.5614 0.9371 0.7362 0.7619
#> [31] 0.7778 0.7061 0.6840 0.9280 0.7700 0.5892 0.6774 0.7712 0.7061 0.8093
#> [41] 0.6427 0.6959 0.7337 0.6628 0.5472 0.4752 0.7163 0.7648 0.6713 0.6054
#> [51] 0.8785 0.6563 0.7386 0.7429 0.8866 0.6626 0.8585 0.7397 0.8182 0.7114
#> [61] 0.7244 0.6841 0.7379 0.5877 0.6097 0.6647 0.5365 0.8555 0.7062 0.4859
#> [71] 0.7458 0.7188 0.6794 0.4648 0.9737
#>