Quantities of Interest from Ecological Inference Estimation

eiread is the command that pulls quantities of interest from the ei object. The command returns a list of quantities of interest requested by the user.

eiread(ei.object, ...)

Arguments

ei.object
An ei object from the function ei.

...
A list of quantities of interest for eiread() to return. See values below.

Value

betab

\(p\) x \(1\) point estimate of \(\beta_i^b\) based on its mean posterior. See section 8.2

betaw

\(p\) x \(1\) point estimate of \(\beta_i^w\) based on its mean posterior. See section 8.2

sbetab

\(p\) x \(1\) standard error for the estimate of \(\beta_i^b\), based on the standard deviation of its posterior. See section 8.2

sbetaw

\(p\) x \(1\) standard error for the estimate of \(\beta_i^w\), based on the standard deviation of its posterior. See section 8.2

phi

Maximum posterior estimates of the CML

psisims

Matrix of random simulations of \(\psi\). See section 8.2

bounds

\(p\) x \(4\): bounds on \(\beta_i^b\) and \(\beta_i^w\), lowerB ~ upperB ~ lowerW ~ upperW. See Chapter 5.

abounds

\(2\) x \(2\): aggregate bounds rows:lower, upper; columns: betab, betaw. See Chapter 5.

aggs

Simulations of district-level quantities of interest \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

maggs

Point estimate of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\) based on the mean of aggs. See Section 8.3.

VCaggs

Variance matrix of 2 district-level parameters, \(\hat{B^b}\) and \(\hat{B^w}\). See Section 8.3.

CI80b

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^b\). See section 8.2.

CI80w

\(p\) x \(2\): lower~upper \(80\%\) confidence intervals for \(\beta_i^w\). See section 8.2.

eaggbias

Regressions of estimated \(\beta_i^b\) and \(\beta_i^w\) on a constant term and \(X_i\).

goodman

Goodman's Regression. See Section 3.1

TODO
References

Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.

Author

Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: molly.e.roberts@gmail.com>>

Examples

data(sample)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample)

#> [1] "Running 2x2 ei"

#> Maximizing likelihood

#> Importance Sampling..

eiread(dbuf, "phi")

#> [1] -1.1534  0.8214 -2.2861 -1.8817  0.4128  0.0000  0.0000

eiread(dbuf, "betab", "betaw")

#> $betab
#>  [1] 0.2040 0.1456 0.1742 0.2864 0.1385 0.1165 0.1016 0.1060 0.2251 0.2545
#> [11] 0.1686 0.2703 0.1649 0.1953 0.1826 0.1952 0.2825 0.1785 0.1516 0.1045
#> [21] 0.3101 0.2555 0.2380 0.1950 0.3913 0.1570 0.1728 0.3959 0.2122 0.2003
#> [31] 0.2329 0.1908 0.1793 0.2596 0.2285 0.1468 0.1898 0.3688 0.2271 0.2579
#> [41] 0.1766 0.1867 0.2014 0.1118 0.1123 0.1178 0.2255 0.2057 0.1453 0.1576
#> [51] 0.2737 0.1699 0.2299 0.2114 0.2731 0.1897 0.2401 0.2469 0.2846 0.2041
#> [61] 0.2124 0.1414 0.2325 0.1641 0.1251 0.1658 0.1541 0.3600 0.2249 0.0399
#> [71] 0.2752 0.1572 0.2031 0.1481 0.3241
#> 
#> $betaw
#>  [1] 0.6908 0.6471 0.6260 0.8811 0.5597 0.3802 0.5255 0.5505 0.7839 0.8998
#> [11] 0.5835 0.9246 0.5985 0.7409 0.7637 0.7112 0.7618 0.6803 0.6808 0.6494
#> [21] 0.8903 0.8736 0.7954 0.7178 0.8732 0.6338 0.5660 0.9568 0.7437 0.7645
#> [31] 0.7637 0.6922 0.6949 0.9280 0.7685 0.6003 0.6779 0.8460 0.7388 0.8274
#> [41] 0.6551 0.7082 0.7338 0.6353 0.5499 0.5000 0.7172 0.7688 0.6311 0.6112
#> [51] 0.8685 0.6687 0.7427 0.7456 0.8800 0.6712 0.8536 0.7778 0.8342 0.7110
#> [61] 0.7263 0.6680 0.7526 0.5927 0.6063 0.6661 0.5447 0.8992 0.7236 0.4934
#> [71] 0.7596 0.7048 0.6797 0.4693 0.9699
#>